+++ src/main/java/in/koreatech/koin/domain/owner/controller/OwnerController.java
@@ -38,7 +38,7 @@ public class OwnerController implements OwnerApi {
 
     @GetMapping("/owner")
     public ResponseEntity<OwnerResponse> getOwner(
-        @Auth(permit = {OWNER}) Long ownerId
+        @Auth(permit = {OWNER}) Integer ownerId
     ) {
         OwnerResponse ownerInfo = ownerService.getOwner(ownerId);
         return ResponseEntity.ok().body(ownerInfo);
