+++ src/main/java/in/koreatech/koin/domain/ownershop/controller/OwnerShopController.java
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import in.koreatech.koin.domain.ownershop.dto.CreateEventRequest;
 import in.koreatech.koin.domain.ownershop.dto.ModifyEventRequest;
 import in.koreatech.koin.domain.ownershop.dto.OwnerShopsRequest;
 import in.koreatech.koin.domain.ownershop.dto.OwnerShopsResponse;
-import in.koreatech.koin.domain.ownershop.dto.CreateEventRequest;
 import in.koreatech.koin.domain.ownershop.service.OwnerShopService;
 import in.koreatech.koin.domain.shop.dto.CreateCategoryRequest;
 import in.koreatech.koin.domain.shop.dto.CreateMenuRequest;
@@ -39,7 +39,7 @@ public class OwnerShopController implements OwnerShopApi {
 
     @GetMapping("/owner/shops")
     public ResponseEntity<OwnerShopsResponse> getOwnerShops(
-        @Auth(permit = {OWNER}) Long ownerId
+        @Auth(permit = {OWNER}) Integer ownerId
     ) {
         OwnerShopsResponse ownerShopsResponses = ownerShopService.getOwnerShops(ownerId);
         return ResponseEntity.ok().body(ownerShopsResponses);
@@ -47,7 +47,7 @@ public class OwnerShopController implements OwnerShopApi {
 
     @PostMapping("/owner/shops")
     public ResponseEntity<Void> createOwnerShops(
-        @Auth(permit = {OWNER}) Long ownerId,
+        @Auth(permit = {OWNER}) Integer ownerId,
         @RequestBody @Valid OwnerShopsRequest ownerShopsRequest
     ) {
         ownerShopService.createOwnerShops(ownerId, ownerShopsRequest);
@@ -56,8 +56,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @GetMapping("/owner/shops/{id}")
     public ResponseEntity<ShopResponse> getOwnerShopByShopId(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable Long id
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable Integer id
     ) {
         ShopResponse shopResponse = ownerShopService.getShopByShopId(ownerId, id);
         return ResponseEntity.ok(shopResponse);
@@ -65,8 +65,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @GetMapping("/owner/shops/menus/{menuId}")
     public ResponseEntity<MenuDetailResponse> getMenuByMenuId(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable Long menuId
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable Integer menuId
     ) {
         MenuDetailResponse menuDetailResponse = ownerShopService.getMenuByMenuId(ownerId, menuId);
         return ResponseEntity.ok(menuDetailResponse);
@@ -74,8 +74,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @GetMapping("/owner/shops/menus")
     public ResponseEntity<ShopMenuResponse> getMenus(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @RequestParam("shopId") Long shopId
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @RequestParam("shopId") Integer shopId
     ) {
         ShopMenuResponse shopMenuResponse = ownerShopService.getMenus(shopId, ownerId);
         return ResponseEntity.ok(shopMenuResponse);
@@ -83,8 +83,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @GetMapping("/owner/shops/menus/categories")
     public ResponseEntity<MenuCategoriesResponse> getCategories(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @RequestParam("shopId") Long shopId
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @RequestParam("shopId") Integer shopId
     ) {
         MenuCategoriesResponse menuCategoriesResponse = ownerShopService.getCategories(shopId, ownerId);
         return ResponseEntity.ok(menuCategoriesResponse);
@@ -92,8 +92,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @DeleteMapping("/owner/shops/menus/{menuId}")
     public ResponseEntity<Void> deleteMenuByMenuId(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("menuId") Long menuId
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("menuId") Integer menuId
     ) {
         ownerShopService.deleteMenuByMenuId(ownerId, menuId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
@@ -101,8 +101,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @DeleteMapping("/owner/shops/menus/categories/{categoryId}")
     public ResponseEntity<Void> deleteCategory(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("categoryId") Long categoryId
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("categoryId") Integer categoryId
     ) {
         ownerShopService.deleteCategory(ownerId, categoryId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
@@ -110,8 +110,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @PostMapping("/owner/shops/{id}/menus")
     public ResponseEntity<Void> createMenu(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("id") Long shopId,
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("id") Integer shopId,
         @RequestBody @Valid CreateMenuRequest createMenuRequest
     ) {
         ownerShopService.createMenu(shopId, ownerId, createMenuRequest);
@@ -120,8 +120,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @PostMapping("/owner/shops/{id}/menus/categories")
     public ResponseEntity<Void> createMenuCategory(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("id") Long shopId,
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("id") Integer shopId,
         @RequestBody @Valid CreateCategoryRequest createCategoryRequest
     ) {
         ownerShopService.createMenuCategory(shopId, ownerId, createCategoryRequest);
@@ -130,8 +130,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @PutMapping("/owner/shops/menus/{menuId}")
     public ResponseEntity<Void> modifyMenu(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("menuId") Long menuId,
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("menuId") Integer menuId,
         @RequestBody @Valid ModifyMenuRequest modifyMenuRequest
     ) {
         ownerShopService.modifyMenu(ownerId, menuId, modifyMenuRequest);
@@ -140,8 +140,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @PutMapping("/owner/shops/menus/categories/{categoryId}")
     public ResponseEntity<Void> modifyMenuCategory(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("categoryId") Long categoryId,
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("categoryId") Integer categoryId,
         @RequestBody @Valid ModifyCategoryRequest modifyCategoryRequest
     ) {
         ownerShopService.modifyCategory(ownerId, categoryId, modifyCategoryRequest);
@@ -150,8 +150,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @PutMapping("/owner/shops/{id}")
     public ResponseEntity<Void> modifyOwnerShop(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("id") Long shopId,
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("id") Integer shopId,
         @RequestBody @Valid ModifyShopRequest modifyShopRequest
     ) {
         ownerShopService.modifyShop(ownerId, shopId, modifyShopRequest);
@@ -160,8 +160,8 @@ public class OwnerShopController implements OwnerShopApi {
 
     @PostMapping("/owner/shops/{id}/event")
     public ResponseEntity<Void> createShopEvent(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("id") Long shopId,
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("id") Integer shopId,
         @RequestBody @Valid CreateEventRequest shopEventRequest
     ) {
         ownerShopService.createEvent(ownerId, shopId, shopEventRequest);
@@ -170,9 +170,9 @@ public class OwnerShopController implements OwnerShopApi {
 
     @PutMapping("/owner/shops/{shopId}/event/{eventId}")
     public ResponseEntity<Void> modifyShopEvent(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("shopId") Long shopId,
-        @PathVariable("eventId") Long eventId,
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("shopId") Integer shopId,
+        @PathVariable("eventId") Integer eventId,
         @RequestBody @Valid ModifyEventRequest modifyEventRequest
     ) {
         ownerShopService.modifyEvent(ownerId, shopId, eventId, modifyEventRequest);
@@ -181,9 +181,9 @@ public class OwnerShopController implements OwnerShopApi {
 
     @DeleteMapping("/owner/shops/{shopId}/event/{eventId}")
     public ResponseEntity<Void> deleteShopEvent(
-        @Auth(permit = {OWNER}) Long ownerId,
-        @PathVariable("shopId") Long shopId,
-        @PathVariable("eventId") Long eventId
+        @Auth(permit = {OWNER}) Integer ownerId,
+        @PathVariable("shopId") Integer shopId,
+        @PathVariable("eventId") Integer eventId
     ) {
         ownerShopService.deleteEvent(ownerId, shopId, eventId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
