+++ src/main/java/in/koreatech/koin/domain/shop/repository/MenuDetailRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.Repository;
 
 import in.koreatech.koin.domain.shop.model.MenuOption;
 
-public interface MenuDetailRepository extends Repository<MenuOption, Long> {
+public interface MenuDetailRepository extends Repository<MenuOption, Integer> {
 
     MenuOption save(MenuOption menuOption);
 }
