+++ src/main/java/in/koreatech/koin/domain/shop/repository/ShopOpenRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.Repository;
 
 import in.koreatech.koin.domain.shop.model.ShopOpen;
 
-public interface ShopOpenRepository extends Repository<ShopOpen, Long> {
+public interface ShopOpenRepository extends Repository<ShopOpen, Integer> {
 
     ShopOpen save(ShopOpen shopOpen);
 
-    List<ShopOpen> findAllByShopId(Long shopId);
+    List<ShopOpen> findAllByShopId(Integer shopId);
 }
